Fix the core files to work with Arduino
The core files are modified to work with Arduino. The main.c file is
modified to create an Arduino task and the main.cpp file is removed.
The Arduino.h file is modified to comment out the include variant.h
line. The app_cfg.h file is modified to add the ARDUINO_APP_TASK_STK_SIZE
define.

The platform.txt file is modified to remove the -mthumb flag
from the compiler.c.elf.flags line and add the -lnosys flag.

The linker_512.ld file is modified to change the vector.o to vector.S.o
and startup.o to startup.S.o.