Update README.md
This commit updates the README.md file to provide a detailed
installation guide for the TOPST VCP Arduino core. The guide includes
steps to download and install the Arduino IDE, add the TOPST VCP board
to the IDE, select the board, and verify the installation. It also
includes a troubleshooting section and references to the Arduino
Troubleshooting Guide and Arduino Help Center for further assistance.