Update release workflow to include file size and sha256 checksum
This commit updates the release workflow to include the file size of the
release file and the sha256 checksum of the release file. The file size
is printed out to the console and the sha256 checksum is written to a
file called `release-files.tar.bz2.sha256`.

This commit also updates the release workflow to include the
`release-files.tar.bz2.sha256` file in the release.