Add Arduino wiring functions support
This commit adds the Arduino wiring functions to the VCP core. The
wiring functions are the basic functions that are used in the Arduino
IDE to interact with the hardware. The wiring functions are used to
configure the pins as input or output, read and write digital values,
and delay the execution of the program.

The wiring functions are added to the VCP core to make it compatible
with the Arduino IDE. The wiring functions are implemented in the
wiring.h, wiring_constants.h, wiring_digital.h, and wiring_time.c files.